backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-jose
pydantic
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
requests
apscheduler

